Create and use eval_MapAt
